Refactored loan calculation to use Strategy and Factory patterns for better modularity and scalability.
