+++ Bot/Handlers/UpdateHandler.cs
@@ -94,6 +94,6 @@ namespace TelegramBot_Fitz.Bot.Handlers
             }
             return _userStates[chatId];
         }
-        
+
     }
 }
