+++ TelegramBot_Fitz.csproj
@@ -129,13 +129,17 @@
     <Compile Include="Bot\UserState.cs" />
     <Compile Include="Bot\BotService.cs" />
     <Compile Include="Core\CalculatorFactory.cs" />
+    <Compile Include="Core\CompoundInterestStrategy.cs" />
     <Compile Include="Core\FloatingCalculator.cs" />
-    <Compile Include="Core\FixedCalculator.cs" />
+    <Compile Include="Core\FixedRateLoanCalculator.cs" />
+    <Compile Include="Core\IInterestCalculationStrategy.cs" />
     <Compile Include="Core\ILoanCalculator.cs" />
+    <Compile Include="Core\InterestCalculationFactory.cs" />
     <Compile Include="Core\InterestCalculationType.cs" />
     <Compile Include="Core\LoanCalculationResult.cs" />
     <Compile Include="Core\OISCalculationResult.cs" />
     <Compile Include="Core\OISCalculator.cs" />
+    <Compile Include="Core\SimpleInterestStrategy.cs" />
     <Compile Include="Core\YearlyCalculation.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
